moved formatting functions to formatters package
