+++ cmd/config/get_config_section.go
@@ -44,7 +44,7 @@ func NewCmdGetConfigSections(config *cfg.Config) *cobra.Command {
 			}
 
 			if strings.HasPrefix(outputFormat, "go-template") {
-				return utils.ApplyGoTemplate(cmd.OutOrStdout(), outputFormat, goTemplate, cc.Items)
+				return formatters.ApplyGoTemplate(cmd.OutOrStdout(), outputFormat, goTemplate, cc.Items)
 			}
 
 			switch outputFormat {
