+++ cmd/coreinstance/get_core_instance.go
@@ -50,7 +50,7 @@ func NewCmdGetCoreInstances(config *cfg.Config) *cobra.Command {
 			}
 
 			if strings.HasPrefix(outputFormat, "go-template") {
-				return utils.ApplyGoTemplate(cmd.OutOrStdout(), outputFormat, goTemplate, aa.Items)
+				return formatters.ApplyGoTemplate(cmd.OutOrStdout(), outputFormat, goTemplate, aa.Items)
 			}
 
 			switch outputFormat {
