+++ cmd/members/get_members.go
@@ -32,7 +32,7 @@ func NewCmdGetMembers(config *cfg.Config) *cobra.Command {
 			}
 
 			if strings.HasPrefix(outputFormat, "go-template") {
-				return utils.ApplyGoTemplate(cmd.OutOrStdout(), outputFormat, goTemplate, mm.Items)
+				return formatters.ApplyGoTemplate(cmd.OutOrStdout(), outputFormat, goTemplate, mm.Items)
 			}
 
 			switch outputFormat {
