+++ cmd/pipeline/create_pipeline.go
@@ -132,7 +132,7 @@ func NewCmdCreatePipeline(config *cfg.Config) *cobra.Command {
 			}
 
 			if strings.HasPrefix(outputFormat, "go-template") {
-				return utils.ApplyGoTemplate(cmd.OutOrStdout(), outputFormat, goTemplate, a)
+				return formatters.ApplyGoTemplate(cmd.OutOrStdout(), outputFormat, goTemplate, a)
 			}
 
 			switch outputFormat {
