+++ cmd/pipeline/create_pipeline_file.go
@@ -51,7 +51,7 @@ func NewCmdCreatePipelineFile(config *cfg.Config) *cobra.Command {
 			}
 
 			if strings.HasPrefix(outputFormat, "go-template") {
-				return utils.ApplyGoTemplate(cmd.OutOrStdout(), outputFormat, goTemplate, out)
+				return formatters.ApplyGoTemplate(cmd.OutOrStdout(), outputFormat, goTemplate, out)
 			}
 
 			switch outputFormat {
