+++ cmd/pipeline/get_pipeline_config.go
@@ -40,7 +40,7 @@ func NewCmdGetPipelineConfigHistory(config *cfg.Config) *cobra.Command {
 			}
 
 			if strings.HasPrefix(outputFormat, "go-template") {
-				return utils.ApplyGoTemplate(cmd.OutOrStdout(), outputFormat, goTemplate, cc.Items)
+				return formatters.ApplyGoTemplate(cmd.OutOrStdout(), outputFormat, goTemplate, cc.Items)
 			}
 
 			switch outputFormat {
