+++ cmd/pipeline/get_pipeline_file.go
@@ -41,7 +41,7 @@ func NewCmdGetPipelineFiles(config *cfg.Config) *cobra.Command {
 			}
 
 			if strings.HasPrefix(outputFormat, "go-template") {
-				return utils.ApplyGoTemplate(cmd.OutOrStdout(), outputFormat, goTemplate, ff.Items)
+				return formatters.ApplyGoTemplate(cmd.OutOrStdout(), outputFormat, goTemplate, ff.Items)
 			}
 
 			switch outputFormat {
@@ -118,7 +118,7 @@ func NewCmdGetPipelineFile(config *cfg.Config) *cobra.Command {
 			}
 
 			if strings.HasPrefix(outputFormat, "go-template") {
-				return utils.ApplyGoTemplate(cmd.OutOrStdout(), outputFormat, goTemplate, file)
+				return formatters.ApplyGoTemplate(cmd.OutOrStdout(), outputFormat, goTemplate, file)
 			}
 
 			switch outputFormat {
