+++ cmd/pipeline/get_pipeline_secret.go
@@ -41,7 +41,7 @@ func NewCmdGetPipelineSecrets(config *cfg.Config) *cobra.Command {
 			}
 
 			if strings.HasPrefix(outputFormat, "go-template") {
-				return utils.ApplyGoTemplate(cmd.OutOrStdout(), outputFormat, goTemplate, ss.Items)
+				return formatters.ApplyGoTemplate(cmd.OutOrStdout(), outputFormat, goTemplate, ss.Items)
 			}
 
 			switch outputFormat {
