+++ cmd/pipeline/get_pipeline_status.go
@@ -40,7 +40,7 @@ func NewCmdGetPipelineStatusHistory(config *cfg.Config) *cobra.Command {
 			}
 
 			if strings.HasPrefix(outputFormat, "go-template") {
-				return utils.ApplyGoTemplate(cmd.OutOrStdout(), outputFormat, goTemplate, ss.Items)
+				return formatters.ApplyGoTemplate(cmd.OutOrStdout(), outputFormat, goTemplate, ss.Items)
 			}
 
 			switch outputFormat {
