+++ cmd/pipeline/rollout_pipeline.go
@@ -10,7 +10,6 @@ import (
 	"gopkg.in/yaml.v2"
 
 	cloud "github.com/calyptia/api/types"
-	"github.com/calyptia/cli/cmd/utils"
 	"github.com/calyptia/cli/completer"
 	cfg "github.com/calyptia/cli/config"
 	"github.com/calyptia/cli/formatters"
@@ -80,7 +79,7 @@ func NewCmdRolloutPipeline(config *cfg.Config) *cobra.Command {
 
 			if autoCreatePortsFromConfig && len(updated.AddedPorts) != 0 {
 				if strings.HasPrefix(outputFormat, "go-template") {
-					return utils.ApplyGoTemplate(cmd.OutOrStdout(), outputFormat, goTemplate, updated)
+					return formatters.ApplyGoTemplate(cmd.OutOrStdout(), outputFormat, goTemplate, updated)
 				}
 
 				switch outputFormat {
