+++ cmd/resourceprofile/create_resource_profile.go
@@ -101,7 +101,7 @@ func NewCmdCreateResourceProfile(config *cfg.Config) *cobra.Command {
 			}
 
 			if strings.HasPrefix(outputFormat, "go-template") {
-				return utils.ApplyGoTemplate(cmd.OutOrStdout(), outputFormat, goTemplate, rp)
+				return formatters.ApplyGoTemplate(cmd.OutOrStdout(), outputFormat, goTemplate, rp)
 			}
 
 			switch outputFormat {
