+++ cmd/resourceprofile/get_resource_profile.go
@@ -50,7 +50,7 @@ func NewCmdGetResourceProfiles(config *cfg.Config) *cobra.Command {
 			}
 
 			if strings.HasPrefix(outputFormat, "go-template") {
-				return utils.ApplyGoTemplate(cmd.OutOrStdout(), outputFormat, goTemplate, pp.Items)
+				return formatters.ApplyGoTemplate(cmd.OutOrStdout(), outputFormat, goTemplate, pp.Items)
 			}
 
 			switch outputFormat {
