+++ cmd/tracerecord/get_trace_record.go
@@ -49,7 +49,7 @@ func NewCmdGetTraceRecords(config *cfg.Config) *cobra.Command {
 			}
 
 			if strings.HasPrefix(outputFormat, "go-template") {
-				return utils.ApplyGoTemplate(cmd.OutOrStdout(), outputFormat, goTemplate, ss.Items)
+				return formatters.ApplyGoTemplate(cmd.OutOrStdout(), outputFormat, goTemplate, ss.Items)
 			}
 
 			switch outputFormat {
