+++ cmd/tracesession/delete_trace_session.go
@@ -52,7 +52,7 @@ func NewCmdDeleteTraceSession(config *cfg.Config) *cobra.Command {
 			}
 
 			if strings.HasPrefix(outputFormat, "go-template") {
-				return utils.ApplyGoTemplate(cmd.OutOrStdout(), outputFormat, goTemplate, terminated)
+				return formatters.ApplyGoTemplate(cmd.OutOrStdout(), outputFormat, goTemplate, terminated)
 			}
 
 			switch outputFormat {
