+++ cmd/tracesession/get_trace_session.go
@@ -57,7 +57,7 @@ func NewCmdGetTraceSessions(config *cnfg.Config) *cobra.Command {
 			}
 
 			if strings.HasPrefix(outputFormat, "go-template") {
-				return utils.ApplyGoTemplate(cmd.OutOrStdout(), outputFormat, goTemplate, ss.Items)
+				return formatters.ApplyGoTemplate(cmd.OutOrStdout(), outputFormat, goTemplate, ss.Items)
 			}
 
 			switch outputFormat {
@@ -125,7 +125,7 @@ func NewCmdGetTraceSession(config *cnfg.Config) *cobra.Command {
 			}
 
 			if strings.HasPrefix(outputFormat, "go-template") {
-				return utils.ApplyGoTemplate(cmd.OutOrStdout(), outputFormat, goTemplate, session)
+				return formatters.ApplyGoTemplate(cmd.OutOrStdout(), outputFormat, goTemplate, session)
 			}
 
 			switch outputFormat {
